+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.6.0] - 2026-05-28
+
 ### Added
 
 - `gemkeeper setup` now accepts a directory path and finds `Gemfile.lock` inside it.
@@ -81,7 +83,8 @@
 
 - Initial release
 
-[Unreleased]: https://github.com/danhorst/gemkeeper/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/danhorst/gemkeeper/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/danhorst/gemkeeper/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/danhorst/gemkeeper/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/danhorst/gemkeeper/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/danhorst/gemkeeper/compare/0.2.1...0.3.0